Refactored security components
